packages/metrics/requirements.txt
paho-mqtt>=1.5.0
prometheus_client>=0.7.1
schema>=0.6.8